requirements.txt
keyboard==0.13.5
python-dotenv==1.1.0
python-vlc==3.0.21203
yt-dlp==2025.3.31
